requirements.txt
openai==1.55.3
pymongo>=4.7
python-dotenv
streamlit>=1.40.2